requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# MongoDB async driver
motor==3.6.0

# Auth & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config & validation
pydantic-settings==2.5.2
pydantic[email]==2.9.2

# Gemini AI (new unified SDK — replaces deprecated google-generativeai)
google-genai>=0.8.0

# PyTorch (CPU build — no CUDA needed for hackathon)
# For GPU: replace torch/torchvision with CUDA variants from pytorch.org
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0        # image decoding for PyTorch

# Utilities
python-multipart==0.0.12
httpx==0.27.2
